scheduler instrumentation: rename "pid" fields to "tid"
The Linux getpid(), gettid() ABIs present "tid == thread id" and "pid ==
process id".

The internal Linux kernel representation is "pid == thread id" and "tgid
== process id".

Given that we want to expose an ABI to end-users that matches the
vision of user-space (rather than that of the implementation), use "tid"
field names to represent thread ids in the sched events.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>