Perf counter context info needs to be at fixed addresses
Fixes a kernel oops when using perf counter with multiple contexts. The
cpu hotplug notifier callback needs to have the callback notifier block
at a fixed address, but the context array may move as it is expanded.

The reason why we use a context array instead of an array of pointers is
to minimize the amount of memory accesses in the tracing hot path.

Let's special-case perf counters and put them in their own memory region
to account for cpu hotplug requirements.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>