Makefile: do not run depmod manually
The modules_install target of the kernel build infrastructure already
runs depmod properly, taking into account the cross-compilation
case. Therefore, it is useless to re-run depmod here, and also harmful
since it does not work in cross-compilation contexts (it does the
depmod on the modules installed on the host and not the ones being
built for the target).

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>