Fix OOPS caused by reference of config pointer
The config pointer points to data that may vanish across the life-time
of the ring buffer stream references. It's only kept valid for the
duration between the channel create and destroy, but the streams can
keep an extra reference on the channel, and thus need the config to stay
valid.

Perform a copy of the config rather than keeping a pointer to it.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>