Xilinx: ARM: Devcfg and SLCR drivers updated to support reconfiguration.
The devcfg and SLCR drivers have been updated to allow (multiple)
reconfigurations of the Programmable Logic.

The flow of reconfiguration is as follows:

1. (Preload): Called in the devcfg driver's open function via an SLCR
function exposed from the SLCR driver.

- Put the FPGA outputs (PS->PL) into reset (disable AXI interface).
- Disable the level shifters.
- Enable the output level shifters.

2. (Reset PROG_B - rising edge): Called in the devcfg driver's open function.

- Set PCFG_PROG_B to 1.
- Set PCFG_PROG_B to 0.
- Wait for PCFG_INIT_B to go low.
- Set PCFG_PROG_B to 1.
- Wait for PCFG_INIT_B to go high
- Clear the PCFG_DONE interrupt.

3. (Download the bitstream): The devcfg write function uses multiple DMA
transfers to send the bitstream to the PCAP.

4. (Postload): Called in the devcfg driver's release function via an SLCR
function exposed from the SLCR driver.

- Enable the input level shifters.
- Take the FPGA outputs (PS->PL) out of reset (enable AXI interface).

The user initiates a reconfiguration from Linux by redirecting a *byte-swapped
binary of the* bitstream into the devcfg device file:
	cat system_bs.bin > /dev/xdevcfg