Xilinx: ARM: EmacPs: Ignore collisions for half duplex mode
For half duplex mode the collisions were being reported as errors.
For half duplex mode collisions are expected to occur for heavy
traffic and it is not an error case. Only when number of retries
exceeds the limit, then only it should be reported as an error.
This patch fixes the issue.

Signed-off-by: Anirudha <anirudh@xilinx.com>