remoteproc: add debugfs entry to print the firmware version info
This patch comes from:
  git://git.ti.com/processor-sdk/processor-sdk-linux.git

A debugfs entry called "version" is created under the corresponding
remoteproc directory when the firmware image is processed. The
design relies on the firmware file having a special section
containing the version info. The version info has to be stored into
this section in the remote processor image build step.

This is currently implemented for ELF images by storing the version
information in a specific ELF section named ".version".

A read on the entry will print the contents of the version section.
The following shows an example usage:
    cat /d/remoteproc/remoteproc0/version

This is based on a patch developed by Juan Gutierrez on 3.4 vendor
kernel, and has been reworked and ported to current kernel.

Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>
(cherry picked from commit 2cad613323d2716f8da2ad28d4ef4fbe6b8343a1)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>