net: xilinx_emacps: Fix driver issue under heavy Rx traffic
The existing driver misbehaves when subjected to heavy traffic
on the Rx path. Under heavy traffic (with a inter packet gap
in the range of 2-5 microseconds) the Rx path of the driver
becomes unresponsive. Under such heavy traffic load it is
natural for the driver to get overwhelmed on Rx path and drop
some packets (as per IEEE specs it is OK), but it should never
become completely unresponsive after the heavy Rx traffic is
taken off. The issue with the existing driver is, under heavy
Rx load, it starts getting lots of error interrupts (buffer not
available) and it does not process them properly. As a result,
at some point the driver consumes all BDs for Rx and does not
free any ensuring a complete lock up.
This patch fixes it by ensuring that even for error cases there
is a constant effort to free up any available BDs.

Signed-off-by: Anirudha Sarangi <anirudh@xilinx.com>
Acked-by: Michal Simek <michal.simek@xilinx.com>