powerpc/book3e: disable the secondary threads when kexec boot
In the embedded multi-threaded implementation, all the threads in a
core share the TLB. On the kexec boot cpu, it will invalidate all the
TLB entries except the one which is used to boot. This will also
invalidate the 1:1 TLB entry which the secondary threads in the same
core are used to spin. Then cause crash on these threads. So we should
disable all the secondary threads when doing kexec boot.

Due to the above reasons, we definitely can't choose a secondary
thread as kexec boot cpu. Since the boot_cpuid is aways thread0,
just choose it.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Signed-off-by: Bin Jiang <bin.jiang@windriver.com>