xilinx_emacps: Add support for gmii to rgmii
This patch adds support for the gmii to rgmii converter IP in the PL.
This Xilinx IP appears on the mdio interface along with the regular PHY.
It possesses a PHY address to be accessible on the mdio bus. The IP appears
in the device tree under the mdio node. This patch reads the mdio node
and on finding the presence of this converter takes appropriate
action assuming it as a dummy PHY. After the regular PHY autoneg, the
driver writes the spped/duplex settings to the appropriate register
in the dummy PHY to ensure proper gmii to rgmii conversion.

Signed-off-by: Anirudha Sarangi <anirudh@xilinx.com>