net: xemacps: Remove warning on round_rate failure
Assume that we deal with an input clock which is adjusted by different
means than the CCF when clk_round_rate() fails and remove the warning
message.
This allows warning free usage of an EMIO clock source w/o a CCF clock
driver on cost of that we don't detect an actual failing call to
clk_round_rate() which I think is rather unlikely to happen.

Signed-off-by: Soren Brinkmann <soren.brinkmann@xilinx.com>
Tested-by: Anirudha Sarangi <anirudh@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>