FogBugz #114479: load custom character for backslash
The NewHaven display displays a Japanese character instead of a
backslash character for ascii 0x5c.  Work around this by loading
a custom backslash character into the display module's cg ram.

Move the command to clear the lcd screen to the last lcd command
in the initialization to clean up.

Signed-off-by: Alan Tull <atull@altera.com>

V2: Added #defines for some constants
[Original patch taken from
 https://github.com/altera-opensource/linux-socfpga.git
 tag rel_socfpga-4.8_17.03.01_pr]
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>