xilinx_emacps: Handle NAPI "rotten packets"
When xemacps_rx_poll() has processed the Rx BD ring until it is empty,
another packet may come in before we have unmasked the Rx interrupts.
No interrupt is signalled for it, and it is left to rot in the BD ring
until the next packet is received.

This is what NAPI calls a "rotting packet", and it actually seems to
occur quite often in heavy traffic (about 30 times per second),
sometimes leading to Rx stalls.

The solution is standard: Check the Rx BD ring again after unmasking
the Rx interrupts, and reschedule NAPI (and remask the interrupts) when
a new packet is detected.

Note that we no longer need an inner loop around xemacps_rx(); the
latter returns when the BD list is empty or the budget is reached, so
we are usually done. [Previously, xemacps_rx() would also return when
XEMACPS_RECV_BD_CNT was reached, so we had to loop over it.]

Signed-off-by: Thomas Betker <thomas.betker@rohde-schwarz.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>