remoteproc: Do not use of_irq_count in xilinx drivers
commit  5d212abb06923d1ac0e33ae1c90f34d9b49f818d from
https://github.com/Xilinx/linux-xlnx.git

Driver resources are already filled and driver can use them.
The reason is that of_irq_count() is not exported
for modules in the mainline.
The next patch which revert this patch which we have in our tree.

Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>