xdevcfg: Automatic endian-swap and header removal
Allows writing a raw .bit file directly to /dev/xdevcfg,
without having to process it using bootgen -split.

The change itself is really quite trivial, but the kernel
implementation is complicated by the fact that userspace
writes the bit-file in small chunks. These chunks may not
be lined up on word boundaries, so it is not possible to
endian-swap the last few bytes, until the next chunk
arrives.

Thus the addition of "residue_buf" to carry across the
last couple of bytes between calls to write() method.

Signed-off-by: Michal Simek <michal.simek@xilinx.com>