fsl_dce: add stream based (de)compression
The fsl_dce_stream object enables the user to do stream based (de)compression.
As opposed to chunk based compression where all the data in a single frame is
sent to the DCE device, stream based (de)compression permits the user to send
mulitple frames and receive the corresponding resuls in multiple frames.

Signed-off-by: Jeffrey Ladouceur <Jeffrey.Ladouceur@freescale.com>
Change-Id: I83065d4681c512d595724b4f92a731de8a032867
Reviewed-on: http://git.am.freescale.net:8181/4297
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Thorpe Geoff-R01361 <Geoff.Thorpe@freescale.com>
Reviewed-by: Wang Haiying-R54964 <Haiying.Wang@freescale.com>
Reviewed-by: Rivera Jose-B46482 <German.Rivera@freescale.com>
[Original patch taken from QorIQ-SDK-V2.0-20160527-yocto]
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>