FogBugz #163905: Support Denali NAND driver on socfpga platform
This patch addresses a few bugs and features in the Denali NAND driver on the socfpga platform.

First the bugs:

- Reading ONFI parameters would cause a timeout, because the code sent the wrong
commands to the controller/device.  That was fixed, and in the process, code was
added to wait for the appropriate interrupt.

- When using the JFFS2 file system, a bug was discovered in the write OOB
data function, whereby the controller was left in SPARE mode causing
subsequent page reads to fail.  The write OOB function was modified
so that the controller is left in MAIN access mode.

Now the features:

- The Denali NAND controller in the socfpga platform has hardware ECC, and several
status bits are different because of that.  A patch from Jamie Iles <jamie at jamieiles.com>
was used as the basis for supporting hardware ECC in the code.

- The device tree and documentation were modified to support hardware ECC.

- One of our test socfpga boards had a NAND chip which was locked by default.  To support
writing to the chip, the UNLOCK1 and UNLOCK2 commands were implemented.  The nand_unlock
calls were added as well.  (Note, the mtd layer has _lock, _unlock, and is probably a
better place to implement this functionality than the nand driver).

- The mtd_device_register call was replaced with mtd_device_parse_register so that
device tree partitions are setup by the mtd layer.  Example partitions were added
to the device tree.

- In the process of supporting our specific controller configuration, duplicate
definitions of bad block table descriptions were removed from the denali code.  They
also exist in the nand driver code.

Signed-off-by: Graham Moore <grmoore@altera.com>
[Original patch taken from
 https://github.com/altera-opensource/linux-socfpga.git
 tag rel_socfpga-4.8_17.03.01_pr]
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>