drm: xilinx: plane: Add properties for z-order and alpha values
When OSD is available, the z-order and alpha values for each plane
can be configured. This patch exposes those using drm properties for planes,
except for the CRTC's private plane. Applications can configure the z-order
and global alpha values for each plane. The layer priority should be unique,
so the layers are sorted by z-position, and the priorities are reassigned
when needed.

xilinx_drm_plane_set_zpos() and xilinx_drm_plane_set_alpha() can be
used externally by the CRTC driver to configure its private plane.

Signed-off-by: Hyun Kwon <hyunk@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>