net: zynq: gem: Fix an error on the RX frame length field in BD for emacps
When you try to catch error frames the kernel panics.
This is due to the constant XEMACPS_RXBUF_LEN_MASK being defined
to 0x3FFF instead of 0x1FFF. When accepting the error frames, the 0x2000 bit
gets set in the BD, giving then a wrong frame length, which is caught
in the kernel when the frame gets passed to le network layers.

This bug has no effect if we don't try to catch error frames.

Signed-off-by: François LEGAL <devel@thom.fr.eu.org>
Acked-by: Punnaiah Choudary Kalluri <punnaia@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>