ARM: OMAP2+: Extend iommu pdata-quirks to DRA7 IPUs & DSP1
The IOMMUs within the IPU and DSP processor subsystems in
DRA7xx are very similar to those in OMAP4/OMAP5, so extend
the OMAP4 iommu pdata quirks for these MMUs as well.

The DSP processor subsystem in DRA7xx, though, has two MMUs,
one for the processor port and another for an EDMA port. Both
these MMUs share a common reset line, the MMU on the EDMA
port will always be mirror-programmed alongside the primary
MMU, so the pdata quirks for reset is only added for the MMU
associated with the processor port.

Signed-off-by: Suman Anna <s-anna@ti.com>
(cherry picked from commit ad6873802c7158eb465ea6febd045585937952af)