arm64: gicv3: its: Increase FORCE_MAX_ZONEORDER for Cavium ThunderX
This patch comes from Cavium online SDK OCTEONTX-SDK-6.1.0-PR.

BugLink: http://bugs.launchpad.net/bugs/1558828

In case of ARCH_THUNDER, there is a need to allocate the GICv3 ITS table
which is bigger than the allowed max order. So we are forcing it only in
case of 4KB page size.

Signed-off-by: Radha Mohan Chintakuntla <rchintakuntla@cavium.com>
Signed-off-by: Robert Richter <rrichter@cavium.com>
[ dannf: Depend on ARM64_4K_PAGES instead of !ARM64_64K_PAGES now that
  16K pages are available ]
Signed-off-by: dann frazier <dann.frazier@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>