xilinx_emacps.c: Attach gmii2rgmii before connecting to actual PHY
Connecting to the PHY starts the state machine and may trigger the
callback. The callback requires the gmii2rgmii interface if that is
being used, so the gmii2rgmii phy must be initialized first.

Instead of "of_phy_connect" use "of_phy_attach", because no state
machine should be started from this phy and it is only used to
communicate settings from the real PHY to the gmii2rgmii logic.

Signed-off-by: Mike Looijmans <mike.looijmans@topic.nl>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>