drm: xilinx: crtc/plane: Attach properties to all planes
This patch attaches the properties to all planes. With the universal
plane, all planes can be exposed to user level, so there's no need
to attach properties to CRTC for private planes. Thus the property
related functions are removed from CRTC, and the properties are
attached to all planes equally.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>