kvm: ignore writing to read-only MSRs
We occationally see this message when we run a guest on OVP host
on Canoe Pass, but it should also be on Nehalem and other Intel
targets too:

[91234.028940] kvm [3889]: vcpu0 unhandled wrmsr: 0x680 data 0
[91234.033709] kvm [3889]: vcpu0 unhandled wrmsr: 0x6c0 data 0
[91234.038477] kvm [3889]: vcpu0 unhandled wrmsr: 0x681 data 0
[91234.043248] kvm [3889]: vcpu0 unhandled wrmsr: 0x6c1 data 0
[91234.048016] kvm [3889]: vcpu0 unhandled wrmsr: 0x682 data 0
[91234.052783] kvm [3889]: vcpu0 unhandled wrmsr: 0x6c2 data 0
[91234.057552] kvm [3889]: vcpu0 unhandled wrmsr: 0x683 data 0
[91234.062320] kvm [3889]: vcpu0 unhandled wrmsr: 0x6c3 data 0
[91234.067088] kvm [3889]: vcpu0 unhandled wrmsr: 0x684 data 0
[91234.071857] kvm [3889]: vcpu0 unhandled wrmsr: 0x6c4 data 0

We see these warnings because kvm_set_msr_common does not contain
handler for the MSR offsets 0x680-0x68f and 0x6c0-0x6cf. According
to this document section 3.8.3:
https://software.intel.com/sites/default/files/76/87/30320
registers offsets 0x680-0x68f and 0x6c0-0x6cf is on-chip LBR stack
containing 16 pairs of MSRs containing source and target addresses of
LBR instructions. These registers are read-only, so we ignore the write
requests to these registers.

Signed-off-by: Nam Ninh <nam.ninh@windriver.com>