kvm: handle read request of perf capabilities from MSR
On Canoe Pass, we occationally see this warning when we run a guest
on OVP host:
[91234.021762] kvm [3889]: vcpu0 unhandled rdmsr: 0x345

This request is to read the perf capabilities from MSR offset 0x345,
but the MSR read handler does not handle it, so we see the warning.
This patch simply reads the KVM native capabilities and returns it
to the guest.

Signed-off-by: Nam Ninh <nam.ninh@windriver.com>