irqchip: Add Xilinx Interrupt Controller driver support
The Xilinx Interrupt Controller IP Core is supported in the MicroBlaze
arch right now and this adds support more generically such that
it can be used with Zynq as a chained (cascaded) controller.

Testing has been done with Zynq only at this point and support
for MicroBlaze is not being moved out of arch/microblaze yet.

Signed-off-by: John Linn <john.linn@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>