FogBugz #261262: Cleanup of SPI device tree (3.18)
When working on FogBugz# 261061, the spi-dev node was moved
into socfpga_cyclone5_socdk.dts as suggested by the upstream
community. These changes were made in 3.17 so to maintain
consistency, this commit brings the same changes to the 3.18
branch. Also noticed that a number of the spi1 fields were
wrong so those were fixed as well.

Signed-off-by: Thor Thayer <tthayer@opensource.altera.com>
[Original patch taken from
 https://github.com/altera-opensource/linux-socfpga.git
 tag rel_socfpga-4.8_17.03.01_pr]
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>