cpu: make clear_tasks_mm_cpumask() lockless when iterating each thread
Because the caller of the function is running offline and
in atomic context, it's not possible to acquire task lock.

It's confirmed that getting rid of the lock is safe due to
the same reason as above. The reader of mm_cpumask() will
not mess up.

There are three kinds of callers of mm_cpumask():
1) switch_mm(); this will not mess up with
   clear_tasks_mm_cpumask() because the two will not run at
   the same time.
2) tlb flush; this will not be affected by
   clear_tasks_mm_cpumask() because clear_tasks_mm_cpumask()
   is called when cpu is offline and the mechanism behind
   tlb flush (which is on_each_cpu) will exclude the
   offlined cpu.
3) Other that check if the current cpu is in mm_cpumask(),
   this doesn't matter because this only happens on the current
   cpu but clear_tasks_mm_cpumask() is called after everything is
   done.

Signed-off-by: Yong Zhang <yong.zhang@windriver.com>
Signed-off-by: Hongbo Zhong <hongbo.zhong@windriver.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>