staging: fsl-mc: Management Complex restool driver
The kernel support for the restool is a driver
for the mc_restool device file. Its purpose is to provide
an ioctl interface, which the restool uses to interact
with the MC bus driver and with the MC firmware.

restool: allocate dpmcp at driver initialization

Current driver does not allocate a dpmcp while initializaing itself.
Current driver only allcoate portal at run time. In current condition,
if dpmcp is used up, restool cannot work at run time.
In order to overcome this defect, we allocate a dpmcp at driver
initialization, and keep that dpmcp until driver exit.
We use that dpmcp by default. If that dpmcp is in use, we create
another portal at run time and destroy the newly created portal after use.

Signed-off-by: Lijun Pan <Lijun.Pan@freescale.com>
[Stuart: squashed commits]
Signed-off-by: Stuart Yoder <stuart.yoder@freescale.com>
[Original patch taken from QorIQ-SDK-V2.0-20160527-yocto]
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>