vfio fsl-mc: Free only inbuilt dprc MC portal
When we bind a DPRC with vfio then we check if it has an
associated initialized MC portal or not. If there is an
initialized MC portal then we destroy and uses vfio allocated
and initialized MC portal. But this will leade to issues
if we unbind the DPRC and then bind again. We should not free
if this is the vfio allocated MC portal.

Signed-off-by: Bharat Bhushan <Bharat.Bhushan@freescale.com>
Change-Id: I054f511b0ec7f656696622db571144a988b01b81
Reviewed-on: http://git.am.freescale.net:8181/35000
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Stuart Yoder <stuart.yoder@freescale.com>
[Original patch taken from QorIQ-SDK-V2.0-20160527-yocto]
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>