staging: fsl_mc: Added udelay() in polling case of mc_send_command()
When calling mc_send_command() from atomic context, we could not call
usleep_range(), so we were just busy waiting, checking the MC status
register in a tight loop. However, this causes a lot of traffic on
the Dickens interconnect from the GPP side, possibly starving the MC.
Thus, we need to have a udelay() in the polling loop of
mc_send_command(), when invoked from atomic context, to throttle down
the traffic on the Dickens interconnect.

Signed-off-by: J. German Rivera <German.Rivera@freescale.com>
Change-Id: I7c2db0cafe5133aacf5b1f535ea31e52ec3584c2
Reviewed-on: http://git.am.freescale.net:8181/35051
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Stuart Yoder <stuart.yoder@freescale.com>
[Original patch taken from QorIQ-SDK-V2.0-20160527-yocto]
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>