dpio/qbman: rename qbman_dq_entry to qbman_result
Currently qbman_dq_entry is used for both dq result in dqrr
and memory, and notifications in dqrr and memory. It doesn't
make sense to have dq_entry in name for those notifications
which have nothing to do with dq. So we rename this as
qbman_result which is meaningful for both cases.

Signed-off-by: Haiying Wang <Haiying.Wang@freescale.com>
Change-Id: I62b3e729c571a1195e8802a9fab3fca97a14eae4
Reviewed-on: http://git.am.freescale.net:8181/35535
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Roy Pledge <roy.pledge@freescale.com>
Reviewed-by: Stuart Yoder <stuart.yoder@freescale.com>
[Original patch taken from QorIQ-SDK-V2.0-20160527-yocto]
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>