staging: fsl-mc: Use msecs_to_jiffies() instead of HZ
Before, we specified the timeout to wait for the completion of an
MC command in jiffies as a fraction of HZ, where HZ is defined as
the number of jiffies in 1 second. However, it is more flexible to
specify this timeout in milliseconds and convert it to jiffies
using msecs_to_jiffies().

Signed-off-by: J. German Rivera <German.Rivera@freescale.com>
Change-Id: I69e2971344a9205ddf2157676b8196ffb43e13fe
Reviewed-on: http://git.am.freescale.net:8181/35644
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Stuart Yoder <stuart.yoder@freescale.com>
[Original patch taken from QorIQ-SDK-V2.0-20160527-yocto]
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>