spi: zynq-qspi: Revise copy_write_data
Combine zynq_qspi_copy_write_data() with writing the TxFIFO register to
simplify the code; rename the function to zynq_qspi_write_tx_fifo().

The new function copies data from xqspi->txbuf and writes it to the
TXD0/TXD1/TXD2/TXD3 register. It also handles the pad byte in the
dual-parallel case, integrating zynq_qspi_tx_dual_parallel().

This is just a cleanup; the control flow is not affected

Signed-off-by: Thomas Betker <thomas.betker@rohde-schwarz.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Reviewed-by: Harini Katakam <harinik@xilinx.com>