spi: zynq-qspi: Handle txlen < 4 in fill_tx_fifo
Merge the code for handling bytes_to_transfer < 4 into
zynq_qspi_fill_tx_fifo(), removing some duplicate code.

Add a parameter 'txempty' for this. Rename the parameter 'size' to
'txcount' as it is a number of words, not a number of bytes.

This is just a cleanup; the control flow is not affected, except that
zynq_qspi_start_transfer() now handles the case transfer->len == 0
correctly.

Signed-off-by: Thomas Betker <thomas.betker@rohde-schwarz.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Reviewed-by: Harini Katakam <harinik@xilinx.com>