Fix: Update kmem event for recent Ubuntu kernel
Ubuntu backported a patch into their 3.13 (Trusty) branch which
changes kmem event definition.

Upstream commit
https://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=99592d598eca62bdbbf62b59941c189176dfc614
first appears in 4.0 (and was later backported to 3.19.2).
We already have the proper check for this.

This corresponds to Ubuntu commit
http://kernel.ubuntu.com/git/ubuntu/ubuntu-trusty.git/commit/?id=d938e4bff1089a1f4757ed94a5cace803c6a363b
which was first released with 3.13.0-50.82.
This required adding a new check specifically for the
Ubuntu version.

Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>