fsl-dpio: Fast DPIO object selection
The DPIO service code had a couple of problems with performance impact:
  - The DPIO service object was protected by a global lock, within
    functions called from the fast datapath on multiple CPUs.
  - The DPIO service code would iterate unnecessarily through its linked
    list, while most of the time it looks for CPU-bound objects.

Add a fast-access array pointing to the same dpaa_io objects as the DPIO
service's linked list, used in non-preemptible contexts.
Avoid list access/reordering if a specific CPU was requested. This
greatly limits contention on the global service lock.
Make explicit calls for per-CPU DPIO service objects if the current
context permits (which is the case on most of the Ethernet fastpath).

These changes incidentally fix a functional problem, too: according to
the specification of struct dpaa_io_notification_ctx, registration should
fail if the specification of 'desired_cpu' cannot be observed. Instead,
dpaa_io_service_register() would keep searching for non-affine DPIO
objects, even when that was not requested.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>
Change-Id: I2dd78bc56179f97d3fd78052a653456e5f89ed82
Reviewed-on: http://git.am.freescale.net:8181/37689
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Roy Pledge <roy.pledge@freescale.com>
Reviewed-by: Stuart Yoder <stuart.yoder@freescale.com>
[Original patch taken from QorIQ-SDK-V2.0-20160527-yocto]
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>