FogBugz #294156: Enable watchdog on Arria10
Enable watchdog timer on Arria10. On the A10, only watchdog1 is able to
trigger a reset reliably.

Also, there is a dependency on the bootloader to enable the boot_clk
source to be from the cb_intosc_hs_clk/2, versus from EOSC1. This
corresponds to the (SWCTRLBTCLKEN | SWCTRLBTCLKSEL) bits enabled
in the clock manager.

Signed-off-by: Dinh Nguyen <dinguyen@opensource.altera.com>
[Original patch taken from
 https://github.com/altera-opensource/linux-socfpga.git
 tag rel_socfpga-4.8_17.03.01_pr]
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>