dpaa2-eth: Add sysfs support for TxConf affinity change
This adds support in sysfs for affining Tx Confirmation queues to GPPs,
via the affine DPIO objects.

The user can specify a cpu list in /sys/class/net/ni<X>/txconf_affinity
to which the Ethernet driver will affine the TxConf FQs, in round-robin
fashion. This is naturally a bit coarse, because there is no "official"
mapping of the transmitting CPUs to Tx Confirmation queues.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>
Change-Id: I4b3da632e202ceeb22986c842d746aafe2a87a81
Reviewed-on: http://git.am.freescale.net:8181/37684
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Stuart Yoder <stuart.yoder@freescale.com>
[Original patch taken from QorIQ-SDK-V2.0-20160527-yocto]
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>