dpaa2-eth: Use CDANs instead of FQDANs
Use Channel Dequeue Available Notifications (CDANs) instead of
Frame Queue notifications. We allocate a QMan channel (or DPCON
object) for each available cpu and assign to it the Rx and Tx conf
queues associated with that cpu.

We usually want to have affine DPIOs and DPCONs (one for each core).
If this is not possible due to insufficient resources, we distribute
all ingress traffic on the cores with affine DPIOs.

NAPI instances are now one per channel instead of one per FQ, as the
interrupt source changes. Statistics counters change accordingly.

Note that after this commit is applied, one needs to provide sufficient
DPCON objects (either through DPL on restool) in order for the Ethernet
interfaces to work.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@freescale.com>
Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>
[Original patch taken from QorIQ-SDK-V2.0-20160527-yocto]
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>