fsl_qman: fix CGR congestion indication processing
QM_MCR_QUERYCONGESTION macro checks if the corresponding bit for
the CGR is set in the "query congestion" command response. If
the corresponding bit is set (meaning that the respective CGR
is in a congested state), then the corresponding callback
is called.
The bit-checking must be done by taking into account the difference
between the DPAA block's and the core's endianness.

Note: improper functioning of CGR callback (i.e. not being called
even if congestion is present) leads to packets being accumulated
in one or more FQs which will eventually lead in the end to
exhausting all the system memory (OOM)

Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
[Original patch taken from QorIQ-SDK-V2.0-20160527-yocto]
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>