remoteproc: zynqmp_r5_remoteproc: release R5 from reset before fw loading
This patch comes from:
  https://github.com/Xilinx/linux-xlnx.git

In case firmware uses TCM, in order to load part of firmware to TCM,
need to release R5 from reset before loading the firmware.

Kernel will not be able to access TCM if it is secure.
There is on going discussion on loading secure firmware, it is not
covered in this implementation.

Signed-off-by: Wendy Liang <jliang@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
(cherry picked from commit 793c5f1a5b773f0cbb3039e17775c61d34ca8e13)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>