ARM: OMAP2+: Extend DRA7 IPU1 MMU pdata quirks to DSP MDMA MMUs
The C66-based DSPs on DRA7xx SoCs do not support a Powerdown-RET
mode, and only supports a Powerdown-Grid OFF mode which requires
a boot from reset. The HW_AUTO setting and a target power domain
state of OFF implies that the DSPs are powered off as soon as
they are idled by executing an IDLE instruction. The DSPs lose
their context as a result and will be unable to resume operations
from any wakeup event.

The DSP power domains therefore need to be restricted to ON state
for the duration a DSP processor is actively running. This is
similar to the restriction required for DRA7 IPU1 processor (albeit
because of a different reason). The IPU1 behavior is handled in
commit 8a8029cc323d ("ARM: OMAP2+: use separate IOMMU pdata to fix
DRA7 IPU1 boot") which adds a .set_pwrdm_constraint ops to the OMAP
IOMMU platform data to restrict the IPU1 power domain to ON state
during the active period of the IPU1 remote processor.

Extend the IPU1 iommu pdata quirks to the DRA7 MDMA MMUs as well
to restrict the DSP power domains to ON state. The MDMA MMU module
configuration will be the first and last steps in the boot and
shutdown sequences of the DSP processors. The existing IPU1 IOMMU
pdata variable has also been renamed appropriately to reflect the
common usage between the IPU1 and the DSPs.

NOTE:
The functionality is inconsistent between different DRA74x and
DRA72x SoCs and silicon revisions. DSP power domains on these
platforms should not be hitting OFF at the moment (even with
firmware images executing an IDLE instruction), because of the
issue described in errata i879 ("DSP MStandby requires CD_EMU in
SW_WKUP") affecting these SoCs. But DSP1 is entering OFF currently
with DRA722 SR1.0, and similar behavior is also seen with DSP1 on
DRA752 SR2.0. DSP1 on DRA722 SR2.0 and DSP2 on DRA752 SR2.0 are
functional for an unexplained reason. The DSP clocks for these
DSPs are gated properly only after the i879 errata fix, while
this patch restricts the DSPs to ON state for proper functionality.

Signed-off-by: Suman Anna <s-anna@ti.com>
(cherry picked from commit 5f95513a4929471fb946d552a560d461b55d2d2e)