MLK-11561-2 regulator: anatop: set default voltage selector for pcie
commit 98e341fc7984db86407160c8b0449175ca198bdf from
git://git.freescale.com/imx/linux-2.6-imx.git imx_4.1.15_1.0.0_ga

The code reads the default voltage selector from its register.
If the default voltage selector is 0 which results in faulty
behaviour of this regulator driver.

This patch sets a default voltage selector for vddpcie-phy if
it is not set in the register.

Signed-off-by: Richard Zhu <Richard.Zhu@freescale.com>