net: phy: Use proper duplex settings for SGMII case
This patch comes from:
  https://github.com/Xilinx/linux-xlnx.git

The duplex setting is not set by the Xilinx phy driver
Which causes issues such that SGMII won't switch rates.
When the duplex is not set it is a value of -1 which then messes
Up the adjust_link() function in the driver as -1 is 0xFFFFFFFF
Which than causes it not to see any link state change.

Reported-by: John Linn <john.linn@xilinx.com>
Signed-off-by: Kedareswara rao Appana <appanad@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
(cherry picked from commit 718784a01910710d2bb3c5aaa5374a1d379a96a8)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>