net: ethernet: xilinx: Fix issues with the optional phy_type property
This patch comes from:
  https://github.com/Xilinx/linux-xlnx.git

The xlnx,phy-type is a optional DT property.
If it is not specified in the device tree then it defaults to 0
Which is the value for XAE_PHY_TYPE_MII which then causes an error
When trying to switch from 10/100 to 1000 when in SGMII mode.

Reported-by: John Linn <john.linn@xilinx.com>
Signed-off-by: Kedareswara rao Appana <appanad@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
(cherry picked from commit e35cbad62bce8710bf74b91b78f4931d83e711de)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>