mtd: nand: Fix oob write operation failures in ddr mode
Found that nand_do_write_oob function issuing the RESET command
to device and causing device to default to its power on state from
the current state.

During the init phase, driver places the device and controller in
to ddr mode and max supported timing mode based on the features
advertised by the device.

So, issuing reset in write_oob operation causes device and controller
in different data interface modes and causing the operation failure.

Signed-off-by: Punnaiah Choudary Kalluri <punnaia@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>