MLK-11676 ARM: imx_v7_defconfig: disable CONFIG_SWP_EMULATE
commit fdc86bf060c834b95119811a76bc61c5444c8dc8 from
git://git.freescale.com/imx/linux-2.6-imx.git imx_4.1.15_1.0.0_ga

According to the help text in the config SWP_EMULATE in arch/arm/mm/Kconfig:

"In some older versions of glibc [<=2.8] SWP is used during futex trylock()
operations with the assumption that the code will not be preempted. This
invalid assumption may be more likely to fail with SWP emulation enabled,
leading to deadlock of the user application."

The audio codec toolchain version is gcc-4.1.1-glibc-2.4, we need turn off
the CONFIG_SWP_EMULATE in the imx_v7_defconfig.

Signed-off-by: Shengjiu Wang <shengjiu.wang@freescale.com>
[Delete modifications about defconfig files]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>