dpaa2-eth: Reduce number of buffers in bpool
Reduce the maximum number of buffers in each buffer pool associated
with a DPNI. This in turn reduces the number of memory allocations
performed in a single batch when buffers fall below a certain
threshold.

Provides a significant performance boost (~5-10% increase) on both
termination and forwarding scenarios, while also reducing the driver
memory footprint.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@freescale.com>
[Original patch taken from QorIQ-SDK-V2.0-20160527-yocto]
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>