remoteproc/keystone: ensure the DSPs are in reset in probe
This patch comes from:
  git://git.ti.com/processor-sdk/processor-sdk-linux.git

The DSPs are expected to be in reset when the driver probes a device.
If the DSPs are out of reset in probe, the system may crash when the
firmware is being loaded. So, add a check to make sure the DSP resets
are asserted, and if not, throw a eye-catchy warning and assert the
resets specifically.

Signed-off-by: Andrew F. Davis <afd@ti.com>
[s-anna@ti.com: replace warning with a WARN]
Signed-off-by: Suman Anna <s-anna@ti.com>

(cherry picked from commit e2f3aed8b4578ab07133b2f47e09d2f4771cb22c)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>