Add support for built-in kernel build
To build the LTTng modules as builtins in a kernel image, simply run
the "built-in.sh" script with the path to you kernel source directory
as an argument. Then configure your kernel build and add the "CONFIG_LTTNG"
option. Build as usual and voila!

This is an early implementation with a single configuration option that
enables or disables everything, We could add fine grained config options
for each modules or group of modules.

Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>