MLK-12103 media: capture: ov5640_mipi: add more delay for camera to stay stable
commit aa7fa01b3ea0e4a970d0970e1cf35701df692b07 from
git://git.freescale.com/imx/linux-2.6-imx.git imx_4.1.15_1.0.0_ga

For power saving, the camera is set ON/OFF by 'PWDN' pin.
However, the DQBUF timeout problems have been observed occasionally when
do the resolution switch for the camera. The reason could be that the power
for the camera is set to off then on when do the switch, but the delay time
is not enough for camera to go stable. The patch sets it to an empirical
value and it has passed over 3-day stress test.

Signed-off-by: Robby Cai <robby.cai@nxp.com>