fsl-dpaa2: eth: Configure a taildrop threshold for each Rx frame queue.
The selected value allows for Rx jumbo (10K) frames processing
while at the same time helps balance the system in the case of
IP forwarding.

Also compute the number of buffers in the pool based on the TD
threshold to avoid starving some of the ingress queues in small
frames, high throughput scenarios.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@freescale.com>
[Original patch taken from QorIQ-SDK-V2.0-20160527-yocto]
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>