fsl-dpaa2: eth: Make DPCON allocation failure produce a benign message
Number of DPCONs may be smaller than the number of CPUs in a number of
valid scenarios. One such scenario is when the DPNI's distribution width
is smaller than the number of cores and we just don't want to
over-allocate DPCONs.
Make the DPCON allocation failure less menacing by changing the logged
message.

While at it, remove a unused parameter in function prototype.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@nxp.com>
[Original patch taken from QorIQ-SDK-V2.0-20160527-yocto]
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>