MLK-12202: ASoC: fsl_asrc_m2m: fix dma timeout after suspend/resume
commit 743cead0f8c4ac6311ffb500efd6146c40124310 from
git://git.freescale.com/imx/linux-2.6-imx.git imx_4.1.15_1.0.0_ga

In suspend function, the complete will be set to done in callback.
After resuming, the convert will not spend time to wait the complete.
which is a wrong complete.
So in suspend function, the complete need to be reinited for next convert.

Signed-off-by: Shengjiu Wang <shengjiu.wang@freescale.com>