powerpc/pm: add the rcpm-wakeup property in wakeup device nodes
Each device with wakeup capability is correspond to one bit in the
IPPDEXPCR (IP Powerdown Exception Control Register). The bit can
determine if the wakeup device is enabled during sleep or deep sleep.

The bit mask is placed in the rcpm-wakeup property of wakeup device
node, such as eSDHC, GPIO, USB1 and USB2.

Signed-off-by: Chenhui Zhao <chenhui.zhao@nxp.com>
[Original patch taken from QorIQ-SDK-V2.0-20160527-yocto]
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>