net: ethernet: xilinx: app4 usage cleanup and tx_skb introduction
This patch comes from:
  https://github.com/Xilinx/linux-xlnx.git

app4 filed is part of DMA BD Structure used by dma
to get the length of the packet in case of rx in case of tx
we are using the same for storing the skb address which is
confusing the user. This patch fixes this
issue by introduing a new tx_skb field.

Signed-off-by: Kedareswara rao Appana <appanad@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
(cherry picked from commit 373401649b471ed33ad4345e3be75f2d874d708f)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>