net: ethernet: xilinx: app4 usage cleanup and tx_skb introduction
app4 filed is part of DMA BD Structure used by dma
to get the length of the packet in case of rx in case of tx
we are using the same for storing the skb address which is
confusing the user. This patch fixes this
issue by introduing a new tx_skb field.

Signed-off-by: Kedareswara rao Appana <appanad@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>