ASoC: Intel: Skylake: Support Pipeline Properties IPC
commit 8f194eabd7ba020c6421d4cd7db05fbeec22e4a8 from
git://git@github.com:01org/linux-intel-4.9.git

Add support to query te properties of any pipeline, that is
running in te ADSP. This ipc involves sending the
pipe id for which properties is sought for.
This IPC is supported in a TLV format and so the
associated changes in the framework are done.
Typical syntax from a debugfs standpoint is
echo 11,<pipeid> > adsp_prop_ctrl

Change-Id: I69f0c5b0a6bbe93587ba6981b0f76e423fb97be8
Signed-off-by: Pardha Saradhi K <pardha.saradhi.kesapragada@intel.com>
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/8241
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Tested-by: Babu, Ramesh <ramesh.babu@intel.com>
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/14046
Reviewed-by: Kale, Sanyog R <sanyog.r.kale@intel.com>
Reviewed-by: Avati, Santosh Kumar <santosh.kumar.avati@intel.com>
Tested-by: Avati, Santosh Kumar <santosh.kumar.avati@intel.com>