soc: qmss: support of loading different types of firmware to pdsp
This patch comes from:
  git://git.ti.com/processor-sdk/processor-sdk-linux.git

Prior to this patch, it only supports the loading of accumulator
firmware to a QMSS PDSP.  This patch adds the support of loading
different types of firmware to a PDSP by specifying a "firmware_type"
in a QMSS PDSP devicetree bindings.

The changes are backward compatible with previous releases in
that if no PDSP firmware_type is specified, it is assumed to
be loaded with accumulator firmware.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Jacob Stiffler <j-stiffler@ti.com>
(cherry picked from commit 7bbeeba6cf4ee9613610914875ddc095fede26b0)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>