staging: fsl-dpaa2/eth: Updates to flow steering support
When flow steering was initially added MC did not support any
masking. Given this and the fact that both hashing and FS must
share a single key, the code was severely limited.

With the addition of masking, flow steering support in ethtool
was somewhat improved. We now take into account the flow type
and also support more key options.

MC currently supports a maximum number of 10 fields in a key.
If and when this limitation is addressed, more flow steering
options can be implemented.

After adding latest FS support code, function prep_cls_rule()
grows to an unmanageable size. So split in separate function for
each user option.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@nxp.com>
[Original patch taken from SDK-V2.0-1703]
Signed-off-by: Meng Li <Meng.Li@windriver.com>