Fix: rename kvm x86 TRACE_SYSTEM to avoid name clash
The lttng-probe-kvm and lttng-probe-kvm-x86 have the same TRACE_SYSTEM
string. Because of that, both modules can't be loaded at the same time.
The second module fails to load silently. Rename the TRACE_SYSTEM of
lttng-probe-kvm-x86 to avoid the clash.

We need to use the LTTNG_TRACEPOINT_EVENT_MAP to perform this
namespacing while still hooking onto the kernel tracepoints.

Fixes #895

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>