Fix: filter interpreter with userspace strings
The filter interpreter is populating the "str" register field, but the
string comparison functions expect to use the "user_str" register field
whenever it deals with a userspace string ("user" flag set to 1).

Since we are using __copy_from_user_inatomic(), this is typically
bailing out on anything that requires to read the string (except the '*'
wildcard).

Fixes #943

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>