ARM: keystone: dts: name netcp clocks for use with runtime_pm
This patch comes from:
  git://git.ti.com/processor-sdk/processor-sdk-linux.git

Netcp driver uses runtime_pm api in probe() and requires clocks to be
named so that it gets enabled. While at it, fix the pa clock to
point to the clkpa which has clock rate of 1/3 of PA PLL clock and add
osr (On Chip Standalone RAM)_clk for K2L as osr is used as external
link ram.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>
(cherry picked from commit ac56953fcaca97382764016c954a85ac8226e480)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>