net:keystone: add support of 10gbe pcsr link status
This patch comes from:
  git://git.ti.com/processor-sdk/processor-sdk-linux.git

The 10GBASE-R Physical Coding Sublayer (PCS-R) module provides
functionality of a physical coding sublayer (PCS) on data being
transferred between a demuxed XGMII and SerDes supporting a 16
or 32 bit interface.  From the driver point of view, whether
a ethernet link is up or not depends also on the status of the
block-lock bit of the PCSR.  This patch adds the checking of that
bit in order to determine the link status.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>
(cherry picked from commit aa46a952af9f689c26a4cd98f45bfd98ed173ee8)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>