net: netcp: remove the redundant memmov()
This patch comes from:
  git://git.ti.com/processor-sdk/processor-sdk-linux.git

The psdata is populated with command data by netcp modules
to the tail of the buffer and knav_dma_set_words() copy the same
to the front of the psdata. So remove the redundant memmov
function call.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>
(cherry picked from commit 798056d54ea5c419bce24042b59370f2fec8d356)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>