ASoC: Intel: Skylake: Additional DSP cores management
commit e5a32e2e30bd58a0c2aaba47d2b8ceba838e31b5 from
git://git@github.com:01org/linux-intel-4.9.git

This patch adds support to manage additional cores on a
need basis. For example, if an algorithm is scheduled to
run on a different core, it would be powered up so
as to enable execution. Once the usecase is complete,
the same would be powered down/usage counter is decremented

Change-Id: I8a195e7b3354bc80031093207d3028e818eaed44
Signed-off-by: Pardha Saradhi K <pardha.saradhi.kesapragada@intel.com>
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/8749
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Tested-by: Babu, Ramesh <ramesh.babu@intel.com>
[Kevin: Just some minor context mods in order to port to wrlinux]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>