powerpc/85xx: drop generic tb sync
These platforms rely on a guts based tb sync
mechanism so no need to default to the generic
tb sync.
This also fixes a kvm over-subscription scenario
where several vcpus of a booting kvm guest share
the same cpu. In such a scenario, the generic
implementation of give_timebase() takes minutes
to finish (~3mins seen on a p5020ds).

Signed-off-by: Laurentiu Tudor <laurentiu.tudor@nxp.com>
[Original patch taken from QorIQ-SDK-V2.0-20160527-yocto]
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>