staging: apf: Making regular use of constant for DMA structure sizes
This patch comes from:
  https://github.com/Xilinx/linux-xlnx.git

Originally a constant was defined to specify sizes of structures
that track DMA state.  This constant was not used, and instead
a literal integer was used when interacting with these structures as
a bound.  This patch makes proper use of the defined constant, and
abandons the literal.

Signed-off-by: Michael Gill <gill@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
(cherry picked from commit 9b663ba03481c4e983895b36bdd8b35f97b04a50)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>