staging: apf: Fixed an error involving invalidation
This patch comes from:
  https://github.com/Xilinx/linux-xlnx.git

If a CPU wrote data to the cache, instructed a DMA to write
to a sub-region an incomplete cache line, then read the line back,
an incorrect data would be read.  This stemmed from how the CPU
invalidated the D-Cache after the DMA finished writing.  The CPU
would invalidate the entire block, assuming the DMA has written the
entire block, thus losing the data that the CPU wrote that the DMA
did not over-write.  This is resolved by manually flushing the value
down, and invalidating before the DMA write is issued, instead of
invalidating after.

Signed-off-by: Michael Gill <gill@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
(cherry picked from commit 3affa58cfc9c9727c0dee7ee12d5738e8ca03777)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>