net: netcp: ale: update to support unknown vlan controls for NU switch
This patch comes from:
  git://git.ti.com/processor-sdk/processor-sdk-linux.git

In NU Ethernet switch used on some of the K2 SoCs, there is separate
UNKNOWNVLAN register for membership, unreg mcast flood, reg mcast flood
and force untag egress bits in ALE. So control for these fields require
different address offset, shift and size of field. This patch adds this
support by checking the revision of the ALE and updating the correct
register info in the control array. The register information available
in ale_controls, needs to be updated to support the netcp NU switch h/w.
So it is not constant array any more since it needs to be updated based
on h/w version. The header of the file is also updated to indicate it
supports N port switch ALE, not just 3 port.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>
(cherry picked from commit 5796c50c3f2f3b36294ee62685491002a92b972e)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>