net: netcp: preparatory patch to introduce sysfs support
This patch comes from:
  git://git.ti.com/processor-sdk/processor-sdk-linux.git

sysfs configuration/control is used for updating the switch
configuration from user space. As there is a significant code
being added for this, it make sense to add it in a separate file
and would require some of the private definitions to be accessed
from the sysfs code. So move common defines to a separate header
file for use in sysfs code.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>
(cherry picked from commit 4b14dea531f240807698615818fd6a31106d7315)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>