crypto/caam: qi - update CGR threshold levels
When the QI is used together with dpa_eth driver, the thresholds
for response queues congestion group needs to take into account
the number of buffers available in the dpa_eth driver. Since the
previous driver version, the kernel configuration define has
changed its name and this patch updates the QI driver accordingly.

Signed-off-by: Alex Porosanu <alexandru.porosanu@nxp.com>
[Original patch taken from QorIQ-SDK-V2.0-20160527-yocto]
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>