net: netcp: ale: add proper ale entry mask bits for netcp switch ALE
This patch comes from:
  git://git.ti.com/processor-sdk/processor-sdk-linux.git

For NetCP NU Switch ALE, some of the mask bits are different than
defaults used in the driver. Add a new macro DEFINE_ALE_FIELD1 that use
a configurable mask bits and use it in the driver. These bits are set to
correct values by using the new variables added to cpsw_ale structure
and re-used in the macros. The parameter nu_switch_ale is configured by
the caller driver to indicate the ALE is for that switch and is used in
the ALE driver to do customization as needed.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>
(cherry picked from commit 397a85ba2a92e5fe7bbe6ff0a5b6ed0904ec7ce0)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>