Documentation: dt: add bindings for ti-cpufreq
This patch comes from:
  git://git.ti.com/processor-sdk/processor-sdk-linux.git

Add the device tree bindings document for the TI CPUFreq/OPP driver
on AM33xx and AM43xx SoCs. This driver is responsible reading and
parsing registers that define which OPPs are available for CPUFreq to
use. From this is sets supported-hw for the OPP framework so that OPPs
can be selectively enabled based on the specific SoC in use.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
(cherry picked from commit cba9c1ca91aad8947a441735e8fab8f32a5437fa)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>