arm64/dts: Add upstream-compatible timer erratum property
The arm,reread-timer property will not be used upstream.  Add
the fsl,erratum-a008585 property which has been proposed upstream,
to ensure that the erratum workaround will be active if an SDK
device tree (or a customer/bootloader fork therefrom) is used
with an upstream kernel (or a future SDK kernel that conforms
with upstream, a distribution kernel, etc).

Signed-off-by: Scott Wood <scott.wood@nxp.com>
[Original patch taken from QorIQ-SDK-V2.0-20160527-yocto]
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>