soc: ti: qos: fix error in misspelled string when parsing dts binding
This patch comes from:
  git://git.ti.com/processor-sdk/processor-sdk-linux.git

The qmss qos clock ticks dts binding is "ticks-per-second".
But the code that parses this binding is wrongly looking for
"tick-per-second".  This results in always using the default
value which happens to be the one supposed to be passed in
from sample devicetrees.  This patch fixes the error.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Jacob Stiffler <j-stiffler@ti.com>
(cherry picked from commit 965a0d216380914cb94fe3f8abd18fde4a1d479e)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>