mmc: host: omap_hsmmc: Fix workaround for errata i834
This patch comes from:
  git://git.ti.com/processor-sdk/processor-sdk-linux.git

The workaround for i834 uses software timeout mechanism for timeouts
that can't be handled by the hardware. But the timeout calculation
assumed the timeout value given by the MMC core (derived from cards
CSD parameters) is for the entire transfer to happen and TC (transfer
complete) interrupt to happen before the timer expires on a successful
transfer. The timeout value actually specifies the delay between the
end bit of the read command and the start bit of the data block and
in the case of multiple-read operation, they also define the typical
delay between the end bit of a data block and the start bit of next
data block (according to the MMC/SD specification). This mismatch
of timeout definition results in the timer expiring before
the TC raising spurious timeouts.

Fix the timeout value to account for the entire transfer to complete
here.
Though the errata i834 is specific to MMCHS HS200 and SDR104,
SW timeout mechanism can be used for any timeout value that can't
be handled by HW.

The disadvantage here is, even if the timeout happens during
the first block, the SW will still wait for expected entire transfer
duration to complete before raising the timeout.

Cc: Mugunthan V N <mugunthanvnm@ti.com>
Cc: Ravikumar Kattekola <rk@ti.com>
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>
(cherry picked from commit d6d077a7bae88770027757cdfccf10b84dd01b95)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>