staging: apf: Fixed DMA-BUF used of buffers spanning partial pages
This patch comes from:
  https://github.com/Xilinx/linux-xlnx.git

The DRM infrastructure when using the xilinx driver has a
potential to return a buffer mapped to a memory region spanning a
partial page, such as in the case of a 1920x1080 resolution
buffer.  When this happens, the scatterlist returned from the DRM
API describes a region rounded up to a whole page.  This patch
trims the returned scatterlist, thus making it usable by a DMA.

Signed-off-by: Michael Gill <gill@xilinx.com>
Signed-off-by: Christian Kohn <christian.kohn@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
(cherry picked from commit ab52f93e8fcb4fb45da9709592fc14f5a37d632b)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>